[REFACTOR] #54 - 패키지 정리
