+++ app/src/main/res/layout/fragment_checklist_registration_dialog.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="com.jjsh.smartshopping.presentation.ui.registration.ChecklistRegistrationViewModel" />
+            type="com.jjsh.smartshopping.presentation.ui.registration.checklist.ChecklistRegistrationViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
